#22 Merged ButtonNamesManager.js into ControllerInfo.js.
